Restructured `generator`
Changed generator to mimic the original structure but avoid lumping all the methods into one file
