+++ hackr/__init__.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-import generator
-import web
-import actions
+from . import generator
+from . import web
+from . import actions
