+++ hackr/generator/internet.py
@@ -1,56 +1,9 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-import random
-import string
-from datetime import datetime
-
+# internet.py
+# Implementation of the faker.provider.internet methods
 from faker import Faker
 
 fake = Faker()
 
-
-def digits(end, count):
-    dataset = []
-    for i in range(count):
-        dataset.append(random.randint(0, end))
-    return dataset
-
-
-def characters(count):
-    dataset = []
-    for i in range(count):
-        dataset.append(random.choice(string.letters))
-    return dataset
-
-
-def dates(count, start_year, end_year):
-    dataset = []
-    for i in range(count):
-        year = random.randint(start_year, end_year)
-        month = random.randint(1, 12)
-        if month != 2:
-            day = random.randint(1, 30)
-        else:
-            day = random.randint(1, 28)
-        dataset.append(str(datetime(year, month, day)))
-    return dataset  # Currently it doesn't include 31st and 29th Feb
-
-
-def names(count):
-    dataset = []
-    for i in range(count):
-        dataset.append(fake.name())
-    return dataset
-
-
-def address(count):
-    dataset = []
-    for i in range(count):
-        dataset.append(fake.address())
-    return dataset
-
-### From faker.providers.internet
 def user_name(count, *args, **kwargs):
     dataset = []
     for i in range(count):
@@ -165,4 +118,3 @@ def uri_extension(count):
     for i in range(count):
         dataset.append(fake.uri_extension())
     return dataset
-### end faker.providers.internet
